@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.16)
 project(RayTracingOneWeekEnd)
 
-set(CMAKE_CXX_STANDARD  17)
+set(CMAKE_CXX_STANDARD  20)
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_SOURCE_DIR}/bin)
 
 add_subdirectory(libs)
